backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
requests
python-dotenv
supabase
langchain-text-splitters
sentence-transformers
pytesseract
pdf2image
Pillow
nest_asyncio